# backend/tests/contracts/snapshots/openapi_v1_snapshot.json
{
  "components": {
    "schemas": {
...
      "KcUserFactRead": {
        "description": "Verified fact response.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fact_type": {
            "title": "Fact Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "valid_from": {
            "format": "date-time",
            "title": "Valid From",
            "type": "string"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "value_json": {
            "title": "Value Json",
            "type": "string"
          },
          "verified_by": {
            "title": "Verified By",
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "fact_type",
          "id",
          "updated_at",
          "user_id",
          "valid_from",
          "value_json",
          "verified_by"
        ],
        "title": "KcUserFactRead",
        "type": "object"
      },
...
      "UserFactRead": {
        "description": "Single fact in GET /user/facts response.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "value_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value Json"
          }
        },
        "required": [
          "confidence",
          "id",
          "key",
          "source",
          "updated_at",
          "user_id"
        ],
        "title": "UserFactRead",
        "type": "object"
      },
      "UserFactUpsertRequest": {
        "additionalProperties": false,
        "description": "POST /user/facts body (authenticated user only; no user_id).",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "number"
              }
            ],
            "default": 0.7,
            "title": "Confidence"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "default": "manual",
            "title": "Source"
          },
          "value_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value Json"
          }
        },
        "required": [
          "key"
        ],
        "title": "UserFactUpsertRequest",
        "type": "object"
      },
...
      "UserProfileKnowledgeRead": {
        "description": "GET /user/knowledge response.",
        "properties": {
          "baseline_summary": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Baseline Summary"
          },
          "constraints_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Constraints Json"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Display Name"
          },
          "goals_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Goals Json"
          },
          "language": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Language"
          },
          "preferences_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Preferences Json"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "updated_at",
          "user_id"
        ],
        "title": "UserProfileKnowledgeRead",
        "type": "object"
      },
      "UserProfileKnowledgeUpsertRequest": {
        "additionalProperties": false,
        "description": "PUT /user/knowledge body (authenticated user only; no user_id).",
        "properties": {
          "baseline_summary": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Baseline Summary"
          },
          "constraints_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Constraints Json"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Display Name"
          },
          "goals_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Goals Json"
          },
          "language": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Language"
          },
          "preferences_json": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ],
            "title": "Preferences Json"
          }
        },
        "title": "UserProfileKnowledgeUpsertRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Sedi is an AI-based health assistant that provides continuous, personalized care. It supports multilingual interaction (English base + Persian + Arabic) and integrates GPT-powered intelligence, adaptive memory, and emotional engagement.",
    "title": "Sedi Intelligent Health Assistant",
    "version": "2.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/knowledge/admin/users/{user_id}/facts": {
      "get": {
        "description": "List verified facts for a user.",
        "operationId": "admin_get_user_facts_knowledge_admin_users__user_id__facts_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KcUserFactRead"
                  },
                  "title": "Response Admin Get User Facts Knowledge Admin Users  User Id  Facts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Get User Facts",
        "tags": [
          "Knowledge Admin"
        ]
      }
    },
...
    "/user/facts": {
      "get": {
        "description": "Get all facts for the authenticated user. Requires Bearer JWT.",
        "operationId": "get_facts_user_facts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserFactRead"
                  },
                  "title": "Response Get Facts User Facts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Facts",
        "tags": [
          "User"
        ]
      },
      "post": {
        "description": "Upsert a fact by key for the authenticated user. Requires Bearer JWT.",
        "operationId": "upsert_fact_user_facts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFactUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Fact",
        "tags": [
          "User"
        ]
      }
    },
    "/user/facts/cleanup": {
      "post": {
        "description": "Admin-only. Find (user_id, key) duplicate groups; keep row with latest updated_at (tie: highest id), delete rest. Requires X-Admin-Token header.",
        "operationId": "cleanup_duplicate_facts_user_facts_cleanup_post",
        "parameters": [
          {
            "description": "If true, only report; if false, delete duplicates",
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "description": "If true, only report; if false, delete duplicates",
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup Duplicate Facts",
        "tags": [
          "User"
        ]
      }
    },
...
    "/user/knowledge": {
      "get": {
        "description": "Get user profile knowledge (baseline). Requires Bearer JWT.",
        "operationId": "get_knowledge_user_knowledge_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileKnowledgeRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Knowledge",
        "tags": [
          "User"
        ]
      },
      "put": {
        "description": "Upsert user profile knowledge (1 row per user). Requires Bearer JWT.",
        "operationId": "upsert_knowledge_user_knowledge_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileKnowledgeUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileKnowledgeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Knowledge",
        "tags": [
          "User"
        ]
      }
    },
...
  }
}